Make client factory per-host and eviction optional
